Consolidated plugins
Related plugins are now grouped together into the same file, improving readability.
